feat: Add 'Why Us' section to CMS config
Integrates a new "Why Us" section into the Netlify CMS configuration. This allows content managers to easily add and edit content for this section, including a title and a list of reasons with associated icons, titles, and descriptions.

Also updates the "steps" widget in the "How it works" section to allow for emojis as well as numbers.